server/requirements.txt
fastapi
uvicorn[standard]
python-multipart
requests
langchain
langchain-core
langchain-google-genai
langgraph
python-dotenv
pydantic
gunicorn
pytest
httpx
weaviate-client
PyYAML